requirements_clean.txt
html2text
datefinder
dask[complete]
nltk
dateparser
sutime
yfinance
pandas
scikit-learn
pandas_market_calendars
pandas 

# Plotting
plotly

# for configuration management
dotmap
pyyaml

# For NGROK-Tunnel
jupyter-server-proxy
pyngrok

# Google Cloud Storage
fsspec==2023.4.0
gcsfs

# For Neural Network Model
transformers
# pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
